Simplify test client setup
